requirements.txt
flask
openai
diffusers
transformers
accelerate
torch
safetensors
Pillow
python-dotenv
flask-mail
